requirements.txt
fastapi>=0.128,<0.129
httpx>=0.28,<0.29
pydantic>=2.12,<3
python-dotenv>=1.1,<2
PyYAML>=6.0,<7
uvicorn>=0.35,<0.36
pytest>=8.4,<9
